arch/arm64: Enable TTL
* For those who don't know about TTL

Time to live (TTL) refers to the amount of time or “hops” that a packet is set to exist inside a network before being discarded by a router. TTL is also used in other contexts including CDN caching and DNS caching.

Reference: https://www.cloudflare.com/learning/cdn/glossary/time-to-live-ttl/

Signed-off-by: karthik558 <karthik.lal558@gmail.com>