arm64: dts: violet: Switch to efficient frequencies
The efficient frequencies are selected from freqbench results [1]
derived from freqbench [2].

Little Clusters:

We are selcting 1593 mhz as the most efficient frequency since this
is the only frequency with the highest iter/mJ value. Also, it took
the least time to complete CoreMark, while just consuming 118mW more
power than the least frequency of the cluster.

Big Cluster:

We are selecting 1209 mhz as the most efficient frequency. Visibly,
it has the highest iter/mJ and lowest energy consumption.

[1] https://github.com/Panchajanya1999/freqbench/blob/master/results/sm6150/main/results.csv
[2] https://github.com/kdrag0n/freqbench

Change-Id: I6c8c1214938f53a6dd4bc250bbbef53ca8d3fc97
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>