drivers/tty: Fixed regression in IRQ
These 2 lines were removed by upstream but it was added back in this particular commit: 49eec58 . Without this fix there will be a memory leak which causes phone to freeze and reboot when su is called by userspace for any usage. This can also cause opening terminal emulators notably termux to freeze and reboot the phone.

Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: K A R T H I K <karthik.lal558@gmail.com>